update for windows compile.

+++ ttLibC/net/tcp.c
@@ -158,7 +158,7 @@ int ttLibC_TcpClient_connect(ttLibC_TcpClientInfo *client_info) {
 			sizeof(addr->addr));
 }
 
-ssize_t ttLibC_TcpClient_read(
+int64_t ttLibC_TcpClient_read(
 		ttLibC_TcpClientInfo *client_info,
 		void * data,
 		size_t data_size) {
